Move components from coin page to home page. Adjust overall layout of home page.
