+++ app/controllers/coins_controller.rb
@@ -10,8 +10,6 @@ class CoinsController < ApplicationController
         .per(100)
         .order(:ranking)
     )
-    @market_dominance = dominance
-    @market_cap = Coin.historical_total_market_data
 
     set_meta_tags(
       title: "Top Cryptocurrency Prices Live, Cryptocurrency Market Cap, Best Cryptocurrency Charts",
@@ -72,26 +70,12 @@ class CoinsController < ApplicationController
     Rollbar.silenced {
       coin_by_id = Coin.find(coin_id)
     }
-    if !coin_by_id 
+    if !coin_by_id
       render_404
     end
 
     # 301 redirect to the same action with the coin slug for SEO purposes
     redirect_to action: action_name, id_or_slug: coin_by_id.slug, status: :moved_permanently
-  end
-
-  def dominance
-    market_dominance = Coin.market_dominance
-
-    # Bitcoin + top 4
-    bitcoin = market_dominance.extract!('bitcoin.org').flat_map { |v| v[1] }
-    other_coins = market_dominance.dup
-                    .sort_by { |k, v| v[:market_percentage] }
-                    .reverse[0..3]
-                    .flat_map { |v| v[1] }
-    coins = bitcoin + other_coins
-
-    coins.as_json
   end
 
   def serialize_coins(coins)
