+++ app/javascript/App/bundles/CoinIndex/index.js
@@ -2,100 +2,13 @@ import React, { Component, Fragment } from 'react'
 import { Table } from 'antd'
 import compose from 'recompose/compose'
 import ColumnNames from './ColumnNames'
-import {
-  Button,
-  Paper,
-  Typography,
-  Hidden,
-  Grid,
-  withWidth,
-} from '@material-ui/core'
+import { withWidth } from '@material-ui/core'
 import SearchCoins from '~/bundles/common/components/SearchCoins'
 import API from '../../bundles/common/utils/localAPI'
 import * as _ from 'lodash'
 import { withStyles, createStyles } from '@material-ui/core/styles'
-import NewsList from './NewsList'
-import MarketDominance from './MarketDominance'
-import TotalMarketCap from './TotalMarketCap'
 
-const styles = (theme) =>
-  createStyles({
-    mainFeaturedPost: {
-      backgroundColor: '#253640',
-      color: '#d7d7d7',
-      boxShadow: 'none',
-    },
-    mainFeaturedPostContent: {
-      padding: `${theme.spacing.unit * 4}px`,
-      [theme.breakpoints.up('md')]: {
-        paddingLeft: 0,
-        paddingRight: 0,
-      },
-    },
-    header: {
-      marginBottom: '0.5em !important',
-      fontSize: '1.5rem',
-      textAlign: 'center',
-    },
-    subheader: {
-      fontSize: '0.8rem',
-      textAlign: 'center',
-      [theme.breakpoints.up('md')]: {
-        fontSize: '1rem',
-      },
-    },
-    cta: {
-      backgroundColor: '#23adf0',
-      color: '#fff',
-      borderRadius: '2px',
-      '&:hover': {
-        backgroundColor: '#23adf0',
-        boxShadow: '0 2px 10px rgba(14, 151, 255, 0.4)',
-      },
-    },
-    widthContainer: {
-      width: '100%',
-      backgroundColor: '#fff',
-    },
-    widgetContainer: {
-      [theme.breakpoints.up('md')]: {
-        margin: '0 auto !important',
-        maxWidth: '1200px',
-        padding: `${theme.spacing.unit * 4}px`,
-      },
-    },
-    widgetContainerLeft: {
-      [theme.breakpoints.down('sm')]: {
-        backgroundColor: 'rgb(7, 29, 41)',
-      },
-      [theme.breakpoints.up('md')]: {
-        maxWidth: '492px',
-      },
-    },
-    widgetContainerRight: {
-      [theme.breakpoints.up('md')]: {
-        height: '100%',
-      },
-    },
-    leftContainerInner: {
-      height: '100%',
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-      alignItems: 'baseline',
-      [theme.breakpoints.up('md')]: {
-        flexDirection: 'column',
-        justifyContent: 'flex-start',
-        alignItems: 'stretch',
-      },
-    },
-    newsWidgetHeader: {
-      [theme.breakpoints.down('sm')]: {
-        borderBottom: '1px solid #e5e8ed',
-        paddingTop: `${theme.spacing.unit * 2}px !important`,
-        paddingBottom: `${theme.spacing.unit}px !important`,
-      },
-    },
-  })
+const styles = (theme) => createStyles({})
 
 class CoinIndex extends Component {
   constructor(props) {
@@ -149,82 +62,8 @@ class CoinIndex extends Component {
     })
   }
   render() {
-    const { classes, marketDominance, totalMarketCap } = this.props
-
     return (
       <Fragment>
-        <Paper className={classes.mainFeaturedPost} square={true}>
-          <Grid
-            container
-            direction="column"
-            justify="center"
-            alignItems="center"
-            className={classes.mainFeaturedPostContent}
-          >
-            <Grid item md>
-              <Typography className={classes.header} color="inherit">
-                Financial intelligence for cryptocurrency
-              </Typography>
-            </Grid>
-            <Grid item md>
-              <Typography
-                className={classes.subheader}
-                color="inherit"
-                paragraph
-              >
-                Uncover buy and sell opportunities through data science backed
-                blockchain analytics.
-              </Typography>
-            </Grid>
-            <Grid item md>
-              <Button
-                className={classes.cta}
-                onClick={() => {
-                  window.location.href = `/signals`
-                }}
-              >
-                Learn more
-              </Button>
-            </Grid>
-          </Grid>
-        </Paper>
-
-        <div className={classes.widthContainer}>
-          <Grid
-            container
-            direction="row"
-            justify="center"
-            alignItems="stretch"
-            className={classes.widgetContainer}
-            spacing={24}
-          >
-            <Grid item xs={12} md={6} className={classes.widgetContainerLeft}>
-              <Grid
-                container
-                className={classes.leftContainerInner}
-                spacing={8}
-              >
-                <Grid item md>
-                  <TotalMarketCap marketCapData={totalMarketCap} />
-                </Grid>
-                <Grid item md>
-                  <MarketDominance coinData={marketDominance} />
-                </Grid>
-              </Grid>
-            </Grid>
-            <Grid item xs={12} md={6} className={classes.widgetContainerRight}>
-              <Typography
-                variant="h5"
-                align="center"
-                className={classes.newsWidgetHeader}
-              >
-                Latest Cryptocurrency News
-              </Typography>
-              <NewsList />
-            </Grid>
-          </Grid>
-        </div>
-
         <div className="flex">
           <h1 className="pt3 pl3">Coins</h1>
           <span style={{ flexGrow: 1 }} />
