+++ app/views/coins/index.slim
@@ -4,7 +4,5 @@
       loggedIn: !!current_user,
       coinCount: @coin_count,
       initialCoins: @coins,
-      marketDominance: @market_dominance,
-      totalMarketCap: @market_cap,
     },
     prerender: true
