+++ app/views/home/index.slim
@@ -4,5 +4,7 @@
       loggedIn: !!current_user,
       coins: @coins,
       watchList: @watched_coins,
+      marketDominance: @market_dominance,
+      totalMarketCap: @market_cap,
     },
     prerender: true
